Modified line 135 to allow more rxs

+++ gprMax/geometry_outputs.py
@@ -132,7 +132,7 @@ class GeometryView(object):
             # Create arrays and add numeric IDs for PML, sources and receivers
             # (0 is not set, 1 is PML, srcs and rxs numbered thereafter)
             self.srcs_pml = np.zeros((G.nx + 1, G.ny + 1, G.nz + 1), dtype=np.int8)
-            self.rxs = np.zeros((G.nx + 1, G.ny + 1, G.nz + 1), dtype=np.int8)
+            self.rxs = np.zeros((G.nx + 1, G.ny + 1, G.nz + 1), dtype=np.int16)
             for pml in G.pmls:
                 self.srcs_pml[pml.xs:pml.xf, pml.ys:pml.yf, pml.zs:pml.zf] = 1
             for index, src in enumerate(G.hertziandipoles + G.magneticdipoles + G.voltagesources + G.transmissionlines):
